isbn adding/editing fix
